requirements-runpod-bulletproof.txt
# Moola RunPod Bulletproof Requirements
# Last Updated: 2025-10-20
# Template: runpod/pytorch:2.4-py3.11-cuda12.4-ubuntu22.04

# CRITICAL: PyTorch Template Compatibility
# DO NOT include torch/torchvision - use template version
# PyTorch 2.4+ supports NumPy 2.0+, avoiding compatibility issues

#############################################
# SYSTEM LIBRARIES (Version Pinning for Stability)
#############################################
# Core numerical computing - Pin to avoid NumPy 2.0 conflicts with older PyTorch
numpy>=1.26.4,<2.1

# Scientific computing
scipy>=1.14,<2.0

#############################################
# DATA PROCESSING LIBRARIES
#############################################
# Data manipulation
pandas>=2.3,<3.0

# High-performance columnar storage (Parquet support)
pyarrow>=17.0,<18.0

#############################################
# MACHINE LEARNING LIBRARIES
#############################################
# Traditional ML
scikit-learn>=1.7,<2.0

# Gradient boosting
xgboost>=2.0,<3.0

# Class imbalance handling (SMOTE)
imbalanced-learn==0.14.0

#############################################
# DEEP LEARNING ECOSYSTEM
#############################################
# High-level PyTorch training framework
pytorch-lightning>=2.4.0,<3.0

# PyTorch metrics library
torchmetrics>=1.8,<2.0

#############################################
# DATA VALIDATION & SCHEMA
#############################################
# Data validation framework
pandera>=0.26.1,<1.0

#############################################
# CONFIGURATION & CLI
#############################################
# Click-based CLI framework
click>=8.2,<9.0

# Modern CLI framework (alternative to click)
typer>=0.17,<1.0

# Configuration management
hydra-core>=1.3,<2.0

# Data validation and settings management
pydantic>=2.11,<3.0
pydantic-settings>=2.9,<3.0

# Environment variable management
python-dotenv>=1.0

#############################################
# LOGGING & OUTPUT
#############################################
# Structured logging
loguru>=0.7,<1.0

# Rich terminal output
rich>=14.0,<15.0

#############################################
# UTILITIES
#############################################
# Progress bars (CRITICAL - missing from current setup)
tqdm>=4.66,<5.0

# Job scheduling
joblib>=1.5,<2.0

# YAML support
PyYAML>=6.0

#############################################
# OPTIONAL AI INTEGRATION
#############################################
# OpenAI integration (future features)
openai>=2.3.0

#############################################
# DEVELOPMENT & TESTING (RunPod Optional)
#############################################
# Uncomment if testing needed on RunPod
# pytest>=8.0,<9.0
# pytest-cov>=4.0
# black>=23.0
# ruff>=0.1
# isort>=5.12

#############################################
# VERSION LOCK RECOMMENDATIONS
#############################################
# For 100% reproducibility, consider exact versions:
# numpy==1.26.4
# pandas==2.3.3
# torch==2.4.0 (from template)
# scipy==1.16.1
# scikit-learn==1.7.2
# xgboost==2.0.3
# pytorch-lightning==2.4.0
# tqdm==4.66.1

#############################################
# KNOWN COMPATIBILITY ISSUES - AVOID
#############################################
# AVOID: torch<2.3 with numpy>=2.0 (incompatible)
# AVOID: Multiple PyTorch installations (use template version)
# AVOID: scikit-learn compiled with different NumPy version
# AVOID: pandas/pyarrow version mismatches